CRUDs actualizados

+++ src/product/product.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, NotFoundException } from '@nestjs/common';
 import { ProductService } from './product.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
@@ -9,6 +9,7 @@ import {
   ApiResponse,
   ApiTags,
   } from '@nestjs/swagger';
+import { ProductEntity } from './entities/product.entity';
 
 @ApiTags('Products api controller')
 @Controller('v1/product')
@@ -42,13 +43,19 @@ export class ProductController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
-    return this.productService.update(+id, updateProductDto);
+  async  update(@Param('id') id: number, @Body() updateProductDto: UpdateProductDto,
+): Promise<ProductEntity>  {
+  const updatedProduct = await this.productService.update(id, updateProductDto);
+  if (!updatedProduct) {
+    throw new NotFoundException(`Producto con ID ${id} no encontrado`);
+  }
+  return updatedProduct;
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.productService.remove(+id);
+  async remove(@Param('id') id: string) {
+    await this.productService.remove(+id)
+    return "Producto borrado";
   }
 
   @Delete('sales/:id')
